Move grid code to own domain
