+++ domains/grid/components/GridWidgetText.vue
