+++ domains/grid/utils/gridConfig.ts
@@ -3,7 +3,7 @@
 
 import Parse from 'parse'
 
-import type { GridWidget, LSP27TheGrid } from '../types/grid'
+import type { GridWidget, LSP27TheGrid } from '../../../types/grid'
 
 const CLASS_NAME = 'grid_config_new'
 
@@ -24,9 +24,9 @@ export type UpsertGridConfigResponse = {
   updatedAt?: string
 }
 
-export async function getGridConfig(
+export const getGridConfig = async (
   username: string
-): Promise<GetGridConfigResponse | undefined> {
+): Promise<GetGridConfigResponse | undefined> => {
   const query: Parse.Query = new Parse.Query(CLASS_NAME)
   query.equalTo('username', username.toLowerCase())
 
@@ -46,10 +46,10 @@ export async function getGridConfig(
   }
 }
 
-export async function upsertGridConfig(
+export const upsertGridConfig = async (
   username: string,
   config: LSP27TheGrid
-): Promise<UpsertGridConfigResponse | undefined> {
+): Promise<UpsertGridConfigResponse | undefined> => {
   const query: Parse.Query = new Parse.Query(CLASS_NAME)
   query.equalTo('username', username.toLowerCase())
